+++ reinforcement_learning/examples/test_cpp/test_data_provider.h
@@ -8,26 +8,26 @@
 
 class test_data_provider {
 public:
-  test_data_provider(const std::string& experiment_name, unsigned int threads, unsigned int examples, unsigned int actions, bool _is_float_reward);
+  test_data_provider(const std::string& experiment_name, size_t threads, size_t examples, size_t actions, bool _is_float_reward);
 
-  const char* get_uuid(unsigned int thread_id, unsigned int example_id) const;
-  const char* get_context(unsigned int thread_id, unsigned int example_id) const;
+  const char* get_uuid(size_t thread_id, size_t example_id) const;
+  const char* get_context(size_t thread_id, size_t example_id) const;
 
-  float get_reward(unsigned int thread_id, unsigned int example_id) const;
-  const char* get_reward_json(unsigned int thread_id, unsigned int example_id) const;
+  float get_reward(size_t thread_id, size_t example_id) const;
+  const char* get_reward_json(size_t thread_id, size_t example_id) const;
 
-  bool is_rewarded(unsigned int thread_id, unsigned int example_id) const;
+  bool is_rewarded(size_t thread_id, size_t example_id) const;
 
-  int report_outcome(reinforcement_learning::live_model* rl, unsigned int thread_id, unsigned int example_id, reinforcement_learning::api_status* status) const;
+  int report_outcome(reinforcement_learning::live_model* rl, size_t thread_id, size_t example_id, reinforcement_learning::api_status* status) const;
 
-  void log(unsigned int thread_id, unsigned int example_id, const reinforcement_learning::ranking_response& response, std::ostream& logger) const;
+  void log(size_t thread_id, size_t example_id, const reinforcement_learning::ranking_response& response, std::ostream& logger) const;
 
 private:
-  std::string create_uuid(const std::string& experiment_name, unsigned int thread_id, unsigned int example_id) const;
-  std::string get_action_features(unsigned int count) const;
-  std::string create_features(unsigned int thread_id, unsigned int example_id) const;
+  std::string create_uuid(const std::string& experiment_name, size_t thread_id, size_t example_id) const;
+  std::string get_action_features(size_t count) const;
+  std::string create_features(size_t thread_id, size_t example_id) const;
   std::string create_context_json(const std::string& cntxt, const std::string& action) const;
-  std::string create_json_reward(unsigned int thread_id, unsigned int example_id) const;
+  std::string create_json_reward(size_t thread_id, size_t example_id) const;
 
 private:
   std::vector<std::vector<std::string>> uuids;
